Update operators.csv
Spectrum was listed twice so I removed the duplicate (It was also not successful when I ran the test so it should be marked not safe). Additionally Spectrum is just a brand and Charter Communications is the actual company and is what shows up in DNS/ASN/etc search so it would be more appropriate as the name in the list (additionally Cloudflare's test itself refers to the company as Charter Communications and not Spectrum so for consistency it makes more sense.